Refactor: TutorialManager 리팩토링 및 위치 이동
- initializeSteps 메서드 삭제
- TutorialStep safe subscript 추가
- isCompleted 속성 삭제 및 currentStep 옵셔널 타입으로 변경
